Separate module for SendingMessages in Slack.
